+++ app/cli.py
@@ -1,8 +1,10 @@
 __G__ = "(G)bd249ce4"
-__V__ = "2020.V.02.00"
+__V__ = "2020.V.02.01b"
 
 from .staticanalyzer import StaticAnalyzer
 from .mics.funcs import killpythoncli,killprocessandsubs
+from .queue.mongoqueue import qbjobqueue
+from .queue.mongoworker import qbworker
 from .logger.logger import logstring,verbose,verbose_flag,verbose_timeout,setuplogger
 from cmd import Cmd
 from os import path,listdir
@@ -10,8 +12,9 @@ from argparse import ArgumentParser
 from shlex import split as ssplit
 from requests import get
 from tempfile import NamedTemporaryFile,gettempdir
-from sys import stdout
+from sys import stdout,argv
 from signal import signal,SIGTSTP,SIGINT
+from .mics.connection import additem,updateitem
 
 def ctrlhandler(signum, frame):
     stdout.write("\n")
@@ -21,7 +24,7 @@ def ctrlhandler(signum, frame):
 print("                                                            ")
 print(" _____  __   _  _____        \\   / ______  ______  _____   ")
 print("|_____| | \\  | |_____| |      \\_/   ____/ |______ |_____/")
-print("|     | |  \\_| |     | |_____  |   /_____ |______ |    \\  {}".format(__V__))
+print("|     | |  \\_| |     | |_____  |   /_____ |______ |    \\ {}".format(__V__))
 print("                               |  https://github.com/QeeqBox/Analyzer")
 print("                                                            ")
 
@@ -36,7 +39,7 @@ class QBAnalyzer(Cmd):
     _analyze_parsergroupreq.add_argument('--file', help="path to file or dump")
     _analyze_parsergroupreq.add_argument('--folder', help="path to folder")
     _analyze_parsergroupreq.add_argument('--buffer', help="input buffer")
-    _analyze_parsergroupdef = _analyze_parser.add_argument_group('Analysis arguments')
+    _analyze_parsergroupdef = _analyze_parser.add_argument_group('Analysis switches')
     _analyze_parsergroupdef.add_argument('--behavior',action='store_true', help="check with generic detections", required=False)
     _analyze_parsergroupdef.add_argument('--xref',action='store_true', help="get cross references", required=False)
     _analyze_parsergroupdef.add_argument('--yara',action='store_true', help="analyze with yara module (Disable this for big files)", required=False)
@@ -58,30 +61,40 @@ class QBAnalyzer(Cmd):
     _analyze_parsergroupdef.add_argument('--worldmap',action='store_true', help="add world map to html", required=False)
     _analyze_parsergroupdef.add_argument('--image',action='store_true', help="add similarity image to html", required=False)
     _analyze_parsergroupdef.add_argument('--full',action='store_true', help="analyze using all modules", required=False)
-    _analyze_parsergroupdeb = _analyze_parser.add_argument_group('Force analysis arguments')
+    _analyze_parsergroupdeb = _analyze_parser.add_argument_group('Force analysis switches')
     _analyze_parsergroupdeb.add_argument('--unicode',action='store_true', help="force extracting ascii", required=False)
     _analyze_parsergroupdeb.add_argument('--bigfile',action='store_true', help="force analyze big files", required=False)
-    _analyze_parsergroupdew = _analyze_parser.add_argument_group('Whitelist')
+    _analyze_parsergroupdew = _analyze_parser.add_argument_group('Whitelist switches')
     _analyze_parsergroupdew.add_argument('--w_internal',action='store_true', help="find it in white list by internal name", required=False)
     _analyze_parsergroupdew.add_argument('--w_original',action='store_true', help="find it in white list by original name", required=False)
     _analyze_parsergroupdew.add_argument('--w_hash',action='store_true', help="find it in white list by hash", required=False)
     _analyze_parsergroupdew.add_argument('--w_words',action='store_true', help="check extracted words against whitelist", required=False)
     _analyze_parsergroupdew.add_argument('--w_all',action='store_true', help="find it in white list", required=False)
-    _analyze_parsergroupdeo = _analyze_parser.add_argument_group('Output arguments')
+    _analyze_parsergroupdeo = _analyze_parser.add_argument_group('Output arguments and switches')
     _analyze_parsergroupdeo.add_argument('--output', help="path of output folder", required=False)
     _analyze_parsergroupdeo.add_argument('--html',action='store_true', help="make html record", required=False)
     _analyze_parsergroupdeo.add_argument('--json',action='store_true', help="make json record", required=False)
     _analyze_parsergroupdeo.add_argument('--open',action='store_true', help="open the report in webbroswer", required=False)
+    _analyze_parsergroupded = _analyze_parser.add_argument_group('Database options')
+    _analyze_parsergroupded.add_argument('--db',action='store_true', help="turn on database option", required=False)
 
-    def __init__(self):
+    def __init__(self,mode):
         super(QBAnalyzer, self).__init__()
         try:
+            logstring("Checking updates","Green")
             ver = get("https://raw.githubusercontent.com/qeeqbox/analyzer/master/version")
             if ver.ok and ver.text.strip() != __V__:
                 logstring("New version {} available, please update.. ".format(ver.text.strip()),"Red")
         except:
             logstring("Update failed","Red")
+
         self.san = StaticAnalyzer()
+        if mode == "--silent":
+            qbjobqueue("jobsqueue",True)
+            qbworker("jobsqueue",self.do_analyze,3)
+            killprocessandsubs()
+        else:
+            self.prompt = "(interactive) "
 
     def help_analyze(self):
         self._analyze_parser.print_help()
@@ -94,7 +107,7 @@ class QBAnalyzer(Cmd):
     analyze --buffer "google.com bit.ly" --full --json --print
     '''
         print(example)
-    	
+
     def do_analyze(self,line):
         try:
             parsed = self._analyze_parser.parse_args(ssplit(line))
@@ -113,9 +126,20 @@ class QBAnalyzer(Cmd):
         else:
             logstring("File, Folder or Buffer is missing","Red")
 
+    @verbose(True,verbose_flag,verbose_timeout,"Loading extra plugins")
+    def savetodb(self,data,parsed):
+        if parsed.db:
+            if len(data)>0:
+                _id = additem("tasks","results",data)
+                if _id:
+                    logstring("Results added to db","Green")
+                else:
+                    logstring("Unable to add result to db","Red")
+
     def analyzefile(self,parsed):
         if path.exists(parsed.file) and path.isfile(parsed.file):
-            self.san.analyze(parsed)
+            data = self.san.analyze(parsed)
+            self.savetodb(data,parsed)
         else:
             logstring("Target File/dump is wrong..","Red")
 
@@ -125,7 +149,8 @@ class QBAnalyzer(Cmd):
                 fullpath = path.join(parsed.folder, f)
                 if path.isfile(fullpath):
                     parsed.file = fullpath
-                    self.san.analyze(parsed)
+                    data = self.san.analyze(parsed)
+                    self.savetodb(data,parsed)
         else:
             logstring("Target folder is wrong..","Red")
 
@@ -135,7 +160,8 @@ class QBAnalyzer(Cmd):
             with open(tempname,"w") as tempfile:
                 tempfile.write(parsed.buffer)
             parsed.file = tempname
-            self.san.analyze(parsed)
+            data = self.san.analyze(parsed)
+            self.savetodb(data,parsed)
         else:
             logstring("Target buffer is empty..","Red")
 
@@ -143,4 +169,14 @@ class QBAnalyzer(Cmd):
         exit()
 
 if __name__ == '__main__':
-    QBAnalyzer().cmdloop()
+    if len(argv) == 2:
+        if argv[1] == "--interactive" or argv[1] == "--silent": 
+            QBAnalyzer(argv[1]).cmdloop()
+
+    print("Please choose a mode:")
+    print("--interactive         Run this framework as an application")
+    print("--silent              Run this framework as service (Required an interface for interaction)")
+    print()
+    print("Examples:")
+    print("python3 -m app.cli --local")
+    print("python3 -m app.cli --service\n")